Bioequivalence Study Data Analysis

Analyze bioequivalence study data with industrial strength. Sample size could be determined for various crossover designs, such as 2x2 design, 2x4 design, 4x4 design, Balaam design, Two-sequence dual design, and William design. Basic assumption is that the variable is distributed as a log-normal distribution. This is SAS PROC GLM style. If you want PROC MIXED style, use nlme::lme.

Details

It performs bioequivalency tests for several variables of a 2x2 study in a data file.

Author

Kyun-Seop Bae <k@acr.kr>

References

Chow SC, Liu JP. Design and Analysis of Bioavailability and Bioequivalence Studies. 3rd ed. (2009, ISBN:978-1-58488-668-6)

Hauschke D, Steinijans V, Pigeot I. Bioequivalence Studies in Drug Development. (2007, ISBN:978-0-470-09475-4)

Diletti E, Hauschke D, Steinijans VW. Sample size determination for bioequivalence assessment by means of confidence intervals. Int J Clinical Pharmacol Ther Tox. 1991;29(1):1-8

Examples

write.csv(NCAResult4BE, "temp.csv", quote=FALSE, row.names=FALSE)
print(be2x2("temp.csv", c("AUClast", "Cmax", "Tmax")), na.print="") 

#> $AUClast
#> $AUClast$`Analysis of Variance (log scale)`
#>                          SS DF           MS           F            p
#> SUBJECT        2.875497e+00 32 8.985928e-02 3.183942248 0.0008742828
#> GROUP          1.024607e-01  1 1.024607e-01 1.145416548 0.2927731856
#> SUBJECT(GROUP) 2.773036e+00 31 8.945279e-02 3.169539016 0.0009544080
#> PERIOD         3.027399e-05  1 3.027399e-05 0.001072684 0.9740824428
#> DRUG           3.643467e-02  1 3.643467e-02 1.290972690 0.2645764201
#> ERROR          8.749021e-01 31 2.822265e-02                         
#> TOTAL          3.786834e+00 65                                      
#> 
#> $AUClast$`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                    0.03061507     0.02822265
#> Coefficient of Variation, CV(%)     17.63193968    16.91883011
#> 
#> $AUClast$`Least Square Means (geometric mean)`
#>                 Reference Drug Test Drug
#> Geometric Means       5092.098  4858.245
#> 
#> $AUClast$`90% Confidence Interval of Geometric Mean Ratio (T/R)`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio    0.889436      0.9540753    1.023412
#> 
#> $AUClast$`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size            6                         7
#> 
#> 
#> $Cmax
#> $Cmax$`Analysis of Variance (log scale)`
#>                          SS DF           MS           F          p
#> SUBJECT        2.861492e+00 32 8.942162e-02 2.237604579 0.01367095
#> GROUP          9.735789e-05  1 9.735789e-05 0.001054764 0.97429977
#> SUBJECT(GROUP) 2.861394e+00 31 9.230304e-02 2.309706785 0.01131826
#> PERIOD         4.717497e-03  1 4.717497e-03 0.118046317 0.73348258
#> DRUG           6.837756e-03  1 6.837756e-03 0.171101730 0.68198228
#> ERROR          1.238856e+00 31 3.996310e-02                       
#> TOTAL          4.112258e+00 65                                    
#> 
#> $Cmax$`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                    0.02616997      0.0399631
#> Coefficient of Variation, CV(%)     16.28355371     20.1921690
#> 
#> $Cmax$`Least Square Means (geometric mean)`
#>                 Reference Drug Test Drug
#> Geometric Means       825.5206  808.8778
#> 
#> $Cmax$`90% Confidence Interval of Geometric Mean Ratio (T/R)`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.9013625      0.9798396    1.065149
#> 
#> $Cmax$`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size            8                         8
#> 
#> 
#> $Tmax
#> $Tmax$`Wilcoxon Signed-Rank Test`
#>   p-value 
#> 0.2326894 
#> 
#> $Tmax$`Hodges-Lehmann Estimate`
#>                            Lower Limit Point Estimate Upper Limit
#> 90% Confidence Interval       -0.33000       -0.03500      0.1050
#> 90% Confidence Interval(%)    74.37661       97.28237    108.1529
#> 
#>