Bioequivalence test for a variable of a 2x2 study

It performs conventional bioequivalence test for 2x2 study. Basic assumption is that the variable is distributed as a log-normal distribution. This is SAS PROC GLM style. If you want PROC MIXED style use nlme::lme.

test2x2(bedata, Var)

Arguments

bedata
Data table name. This should have at least the following columns and a variable column to be tested. Var column values should be all positive values.

  
 GRP : Group or Sequence, 'RT' or 'TR'
 PRD : Period, 1 or 2
 SUBJ : Subject ID
 TRT : Treatment or Drug, 'R' or 'T'

Var
Variable to be tested. This should be one of the column names in bedata table. Usually 'AUClast' or 'Cmax'

Details

It tests Var variable equivalency from a 2x2 study. Current regulatory requirement is that the 90% confidence interval of geometric mean ratio (Test/Reference) should be within [0.8, 1.25].

Value

Analysis of Variance (log scale)

Analysis of Variance in log scale

Between and Within Subject Variability

Variance in log scale and coefficient of variance in original scale

Least Square Means

Geometric means

90% Confidence Interval

90% confidence interval of geometric mean ratio (T/R)

Sample Size

Sample size for the replication of this study

Author

Kyun-Seop Bae <k@acr.kr>

Examples

print(test2x2(NCAResult4BE, "AUClast"), na.print="") 

#> $`Analysis of Variance (log scale)`
#>                          SS DF           MS           F            p
#> SUBJECT        2.875497e+00 32 8.985928e-02 3.183942248 0.0008742828
#> GROUP          1.024607e-01  1 1.024607e-01 1.145416548 0.2927731856
#> SUBJECT(GROUP) 2.773036e+00 31 8.945279e-02 3.169539016 0.0009544080
#> PERIOD         3.027399e-05  1 3.027399e-05 0.001072684 0.9740824428
#> DRUG           3.643467e-02  1 3.643467e-02 1.290972690 0.2645764201
#> ERROR          8.749021e-01 31 2.822265e-02                         
#> TOTAL          3.786834e+00 65                                      
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                    0.03061507     0.02822265
#> Coefficient of Variation, CV(%)     17.63193968    16.91883011
#> 
#> $`Least Square Means (geometric mean)`
#>                 Reference Drug Test Drug
#> Geometric Means       5092.098  4858.245
#> 
#> $`90% Confidence Interval of Geometric Mean Ratio (T/R)`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio    0.889436      0.9540753    1.023412
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size            6                         7
#> 

print(test2x2(NCAResult4BE, "Cmax"), na.print="") 

#> $`Analysis of Variance (log scale)`
#>                          SS DF           MS           F          p
#> SUBJECT        2.861492e+00 32 8.942162e-02 2.237604579 0.01367095
#> GROUP          9.735789e-05  1 9.735789e-05 0.001054764 0.97429977
#> SUBJECT(GROUP) 2.861394e+00 31 9.230304e-02 2.309706785 0.01131826
#> PERIOD         4.717497e-03  1 4.717497e-03 0.118046317 0.73348258
#> DRUG           6.837756e-03  1 6.837756e-03 0.171101730 0.68198228
#> ERROR          1.238856e+00 31 3.996310e-02                       
#> TOTAL          4.112258e+00 65                                    
#> 
#> $`Between and Within Subject Variability`
#>                                 Between Subject Within Subject
#> Variance Estimate                    0.02616997      0.0399631
#> Coefficient of Variation, CV(%)     16.28355371     20.1921690
#> 
#> $`Least Square Means (geometric mean)`
#>                 Reference Drug Test Drug
#> Geometric Means       825.5206  808.8778
#> 
#> $`90% Confidence Interval of Geometric Mean Ratio (T/R)`
#>                  Lower Limit Point Estimate Upper Limit
#> 90% CI for Ratio   0.9013625      0.9798396    1.065149
#> 
#> $`Sample Size`
#>                       True Ratio=1 True Ratio=Point Estimate
#> 80% Power Sample Size            8                         8
#>